requirements.txt
pydantic-ai>=0.2.0
pydantic>=2.0
openai>=1.0
httpx>=0.27
psycopg2-binary>=2.9
python-dotenv>=1.0
rich>=13.0
